# geojson: add support for "external" json encoders/decoders